Change guid route to file detail route.
