+++ website-ember/app/controllers/file-detail.js
