+++ website-ember/app/templates/file-detail.hbs
@@ -1,10 +1,5 @@
 <br>
 <br>
-<br>
-<br>
-<br>
-<br>
-
 <div class='row'>
     <div class='col-sm-5'>
         <h2>{{model.name}}</h2>
